finished the code
While in both solutions, I am essentially doing the same thing, in 1 approach, I use a List of List to create by buckets of frequencies, while in the 2nd approach I use an Array of List of Integers. The 2nd approach is smaller code size, but makes it easier to read, at least for me.